test-requirements.txt
docker-compose
mock
nose
pyhamcrest
requests
sh
